created a new folder for add dialog in customerdata, to optimize code and make it easy to understand
