+++ src/components/index.ts
@@ -3,4 +3,5 @@ export * from './NavBar'
 export * from './AddNew'
 export * from './UploadFile'
 export * from './FileViewDialog'
-export * from "./WholeReportDialog"
+export * from "./WholeReportDialog";
+export * from "./AddClientForm"
